+++ components/ast.h
@@ -5,9 +5,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-#define ARENA_BLOCK_SIZE 8192  // 8KB blocks for memory pooling in order to remove memory leaks
+#define ARENA_BLOCK_SIZE 8192 // 8KB blocks for memory pooling in order to remove memory leaks
 
-typedef enum {
+typedef enum
+{
     NODE_NUMBER,
     NODE_BINARY_OP,
     NODE_ASSIGN,
@@ -17,61 +18,89 @@ typedef enum {
     NODE_FUNC_DEF,
     NODE_FUNC_CALL,
     NODE_FOR,
-    NODE_VAR_DECL
+    NODE_VAR_DECL,
+    NODE_LOGICAL_OP,
+    NODE_RELATIONAL_OP,
+    NODE_COMPARISON_OP,
 } NodeType;
 
 // Arena allocator structure
-typedef struct Arena {
+typedef struct Arena
+{
     char *block;
     size_t offset;
     size_t size;
     struct Arena *next;
 } Arena;
 
-typedef struct ASTNode {
+typedef struct ASTNode
+{
     NodeType type;
-    union {
+    union 
+    {
         int number;
-        struct {
+        struct
+        {
             char op;
             struct ASTNode *left, *right;
         } binaryOp;
-        struct {
+        struct
+        {
             char name[50];
             struct ASTNode *expr;
         } assign;
-         struct {
-            char name[50]; 
+        struct
+        {
+            char name[50];
             ASTNode *value;
         } varDecl;
-        struct {
+        struct
+        {
             struct ASTNode *condition;
             struct ASTNode *thenStmt;
             struct ASTNode *elseStmt;
         } ifNode;
-        struct {
+        struct
+        {
             struct ASTNode *condition;
             struct ASTNode *body;
         } whileNode;
-        struct {
+        struct
+        {
             struct ASTNode *condition;
             struct ASTNode *body;
         } doWhileNode;
-        struct {
+        struct
+        {
             char name[50];
             struct ASTNode *params;
             struct ASTNode *body;
         } funcDef;
-        struct {
+        struct
+        {
             char name[50];
             struct ASTNode *args;
         } funcCall;
-        struct {
+        struct
+        {
             struct ASTNode *initialization;
             struct ASTNode *condition;
             struct ASTNode *increment;
             struct ASTNode *body;
         } forNode;
+        struct
+        {
+            char op[3];
+            struct ASTNode *left, *right;
+        } relOp;
+
+        struct
+        {
+            char op[3];
+            struct ASTNode *left, *right;
+        } logicalOp;
+
+        int boolean;
     };
     struct ASTNode *next;
 } ASTNode;
@@ -84,11 +113,14 @@ ASTNode *astTail = NULL;
 Arena *arenaHead = NULL;
 
 // Function to allocate a new memory block in the arena
-void *arenaAlloc(size_t size) {
-    if (!arenaHead || arenaHead->offset + size > arenaHead->size) {
+void *arenaAlloc(size_t size)
+{
+    if (!arenaHead || arenaHead->offset + size > arenaHead->size)
+    {
         size_t blockSize = (size > ARENA_BLOCK_SIZE) ? size : ARENA_BLOCK_SIZE;
         Arena *newBlock = (Arena *)malloc(sizeof(Arena) + blockSize);
-        if (!newBlock) {
+        if (!newBlock)
+        {
             fprintf(stderr, "Memory allocation failed\n");
             exit(1);
         }
@@ -104,15 +136,18 @@ void *arenaAlloc(size_t size) {
 }
 
 // Function to allocate an ASTNode using the arena
-ASTNode *allocateNode(NodeType type) {
+ASTNode *allocateNode(NodeType type)
+{
     ASTNode *node = (ASTNode *)arenaAlloc(sizeof(ASTNode));
     memset(node, 0, sizeof(ASTNode));
     node->type = type;
     return node;
 }
 
-void freeArena() {
-    while (arenaHead) {
+void freeArena()
+{
+    while (arenaHead)
+    {
         Arena *temp = arenaHead;
         arenaHead = arenaHead->next;
         free(temp);
@@ -120,3 +155,4 @@ void freeArena() {
 }
 
 #endif
+