+++ changes.md
@@ -1,4 +1,14 @@
 
+## 28 Mar 2025
+- added semantic analysis
+- added code generation
+- added Memory Management using Arena Allocator
+- added Precedence Climbing for Expression Parsing
+
+Creator [Jeevan Joshi](https://github.com/jeevanjoshi4434)
+
+---
+
 ## 15 Feb 2025
 
 - added lexical analysis using `flex` and `bison`